Fixes #10087: Correct display of far end in console/power/interface connections tables
